Tweaks to XCM for Benchmarking (#4283)
* tweaks to xcm stuff for benchmarking

* Update xcm/xcm-executor/src/lib.rs
